pyproject.toml
[tool.poetry]
name = "the-office"
version = "0.1.0"
description = ""
authors = ["Lalith Sagar  Devagudi <lalithsagardevagudi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-openai = "^0.3.3"
python-dotenv = "^1.1.0"
langchain = "^0.3.24"
langchain-community = "^0.3.22"
uvicorn = "^0.34.2"
fastapi = "^0.115.12"
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"
llama-index-core = "^0.12.41"
langgraph = "^0.4.8"
openai = "^1.86.0"
serpapi = "^0.1.5"
langchain-core = "^0.3.65"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
